pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "stroke-eye-monitor"
version = "0.1.0"
description = "Webcam eye/iris tracking and optional gaze calibration with MediaPipe Face Landmarker"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
    "mediapipe>=0.10.9,<0.11",
    "opencv-python>=4.8.0,<5",
    "numpy>=1.24.0,<3",
    "certifi>=2024.0.0",
    "openai>=1.40.0,<2",
    "python-dotenv>=1.0.0,<2",
    "sounddevice>=0.4.6,<1",
]

[project.optional-dependencies]
audio = [
    "soundcard>=0.4.2,<1",
]

[project.scripts]
stroke-eye-monitor = "stroke_eye_monitor.app:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = [
  "src"
]